[Refactor] #1301 Renamed apply-magic-armors-base.cpp/h to abstract-protector-enchanter.cpp/h
