+++ Hengband/Hengband/Hengband.vcxproj
@@ -273,7 +273,7 @@
     <ClCompile Include="..\..\src\main-win\main-win-term.cpp" />
     <ClCompile Include="..\..\src\main-win\wav-reader.cpp" />
     <ClCompile Include="..\..\src\object-enchant\apply-magic-amulet.cpp" />
-    <ClCompile Include="..\..\src\object-enchant\apply-magic-armors-base.cpp" />
+    <ClCompile Include="..\..\src\object-enchant\abstract-protector-enchanter.cpp" />
     <ClCompile Include="..\..\src\object-enchant\apply-magic-boots.cpp" />
     <ClCompile Include="..\..\src\object-enchant\apply-magic-cloak.cpp" />
     <ClCompile Include="..\..\src\object-enchant\apply-magic-crown.cpp" />
@@ -959,7 +959,7 @@
     <ClInclude Include="..\..\src\main-win\main-win-term.h" />
     <ClInclude Include="..\..\src\main-win\wav-reader.h" />
     <ClInclude Include="..\..\src\object-enchant\apply-magic-amulet.h" />
-    <ClInclude Include="..\..\src\object-enchant\apply-magic-armors-base.h" />
+    <ClInclude Include="..\..\src\object-enchant\abstract-protector-enchanter.h" />
     <ClInclude Include="..\..\src\object-enchant\apply-magic-boots.h" />
     <ClInclude Include="..\..\src\object-enchant\apply-magic-cloak.h" />
     <ClInclude Include="..\..\src\object-enchant\apply-magic-crown.h" />
