+++ src/Makefile.am
@@ -566,11 +566,11 @@ hengband_SOURCES = \
 	object-activation/activation-teleport.cpp object-activation/activation-teleport.h \
 	object-activation/activation-util.cpp object-activation/activation-util.h \
 	\
+	object-enchant/abstract-protector-enchanter.cpp object-enchant/abstract-protector-enchanter.h \
 	object-enchant/activation-info-table.cpp object-enchant/activation-info-table.h \
 	object-enchant/apply-magic.cpp object-enchant/apply-magic.h \
 	object-enchant/apply-magic-amulet.cpp object-enchant/apply-magic-amulet.h \
 	object-enchant/apply-magic-armor.cpp object-enchant/apply-magic-armor.h \
-	object-enchant/apply-magic-armors-base.cpp object-enchant/apply-magic-armors-base.h \
 	object-enchant/apply-magic-boots.cpp object-enchant/apply-magic-boots.h \
 	object-enchant/apply-magic-cloak.cpp object-enchant/apply-magic-cloak.h \
 	object-enchant/apply-magic-crown.cpp object-enchant/apply-magic-crown.h \
