+++ src/object-enchant/abstract-protector-enchanter.h
