+++ src/object-enchant/apply-magic-armor.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 
 #include "object-enchant/enchanter-base.h"
-#include "object-enchant/apply-magic-armors-base.h"
+#include "object-enchant/abstract-protector-enchanter.h"
 #include "system/angband.h"
 
 struct object_type;
